ml/requirements.txt
torch>=2.5.0
transformers>=4.46.0
datasets>=3.1.0
evaluate>=0.4.3
peft>=0.13.2
optimum[onnxruntime]>=1.23.3
trl>=0.12.1
wandb>=0.18.7
optuna>=4.1.0
pyyaml>=6.0.2
